[AL-10018] Split tests for clients
