+++ tests/test_client.py
@@ -1,11 +1,10 @@
 import unittest
 import requests
-import aiohttp
 
-from unittest.mock import MagicMock, AsyncMock
+from unittest.mock import MagicMock
 from sdc_client import client
 from sdc_client.base_api_client import UnauthorizedException
-from conftest import MockResponse, MockAsyncResponse, StreamSetsMock
+from conftest import MockResponse, StreamSetsMock
 
 
 class TestGetJMX(unittest.TestCase):
@@ -24,30 +23,6 @@ class TestGetJMX(unittest.TestCase):
         ))
         with self.assertRaises(UnauthorizedException):
             client.get_jmx(StreamSetsMock(), 'query_params_1')
-
-    def test_get_jmx_async_success(self):
-        aiohttp.ClientSession.get = AsyncMock(return_value=MockAsyncResponse(
-            _text='data',
-            status_code=200
-        ))
-        queries = [
-            (StreamSetsMock(), 'query_params_1',),
-            (StreamSetsMock(), 'query_params_2',),
-        ]
-        result = client.get_jmxes_async(queries=queries)
-        self.assertEqual(result, [{'json': 'data'}, {'json': 'data'}])
-
-    def test_get_jmx_async_unauthorized(self):
-        aiohttp.ClientSession.get = AsyncMock(return_value=MockAsyncResponse(
-            _text='data',
-            status_code=401,
-        ))
-        queries = [
-            (StreamSetsMock(), 'query_params_1',),
-            (StreamSetsMock(), 'query_params_2',),
-        ]
-        with self.assertRaises(UnauthorizedException):
-            client.get_jmxes_async(queries=queries)
 
 
 if __name__ == '__main__':
